pyproject.toml
[project]
name = "python-intermedio-pep8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
]
